requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.27
openai==1.25.0
python-arango>=7.9,<8
sentence-transformers==2.2
websockets==12.0
influxdb-client==1.40
cachetools==5.3
colorama==0.4
pdfminer.six==20221105
structlog>=24,<25
pydantic==2.7
